# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Proyectos de Pavimento",
        "description": "Documentación de la API para la gestión de proyectos, abscisas, losas y patologías. Creada para la aplicación móvil y web.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/abscisa": {
            "get": {
                "tags": [
                    "Abscisas"
                ],
                "summary": "Obtener lista paginada de abscisas por proyecto",
                "operationId": "401231738edccd7c098223b1506fecf8",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "query",
                        "description": "ID del proyecto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "description": "Objeto de paginación de Laravel",
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Abscisa"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Parámetros inválidos"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/abscisa/create": {
            "post": {
                "tags": [
                    "Abscisas"
                ],
                "summary": "Crear una nueva abscisa",
                "operationId": "15a0a23199a0916b02cddba83de192f3",
                "requestBody": {
                    "description": "Datos de la nueva abscisa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "number_of_abscisas",
                                    "project_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "K1+000 - K2+000"
                                    },
                                    "number_of_abscisas": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "project_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Abscisa creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Abscisa"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/abscisa/delete": {
            "delete": {
                "tags": [
                    "Abscisas"
                ],
                "summary": "Eliminar una abscisa",
                "operationId": "d794084aa0a93dd1352325ef9c11aaca",
                "parameters": [
                    {
                        "name": "abscisa",
                        "in": "path",
                        "description": "ID de la abscisa a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Abscisa eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Abscisa no encontrada"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/web/projects/show": {
            "get": {
                "tags": [
                    "Abscisas",
                    "Projects"
                ],
                "summary": "Obtener todas las abscisas de un proyecto con sus losas",
                "operationId": "6a33a8833e35d8951d6d2488e80046f6",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "description": "ID del proyecto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Abscisa"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/api/abscisa/chart/show/{abscisa_id}": {
            "get": {
                "tags": [
                    "Abscisas"
                ],
                "summary": "Obtener datos de patologías y daños para una abscisa",
                "operationId": "8def88af18f042c0c20b60a13d55c1f4",
                "parameters": [
                    {
                        "name": "abscisa",
                        "in": "path",
                        "description": "ID de la abscisa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "Abscisa no encontrada"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/api/abscisa/change/status": {
            "patch": {
                "tags": [
                    "Abscisas"
                ],
                "summary": "Actualizar el estado de una abscisa",
                "operationId": "d0b3208849258ac4d16fa6d04459687c",
                "parameters": [
                    {
                        "name": "abscisa",
                        "in": "path",
                        "description": "ID de la abscisa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Nuevo estado de la abscisa",
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "in_progress",
                                            "finished"
                                        ],
                                        "example": "finished"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado actualizado exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "404": {
                        "description": "Abscisa no encontrada"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Abscisa": {
                "title": "Abscisa",
                "description": "Modelo de Abscisa. Representa una sección o tramo específico dentro de un proyecto, el cual contiene múltiples losas (placas).",
                "properties": {
                    "id": {
                        "description": "El ID único de la abscisa",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre o identificador de la abscisa (ej. K0+000)",
                        "type": "string",
                        "example": "K0+000 - K1+000"
                    },
                    "number_of_abscisas": {
                        "description": "Número de mediciones o sub-secciones dentro de este tramo",
                        "type": "integer",
                        "example": 100
                    },
                    "project_id": {
                        "description": "ID del proyecto al que pertenece",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "ID del usuario que la registró",
                        "type": "integer",
                        "example": 12
                    },
                    "status": {
                        "description": "Estado actual de la abscisa",
                        "type": "string",
                        "enum": [
                            "pending",
                            "in_progress",
                            "finished"
                        ],
                        "example": "pending"
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-27T10:00:00Z"
                    },
                    "updated_at": {
                        "description": "Fecha y hora de la última actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-27T12:30:00Z"
                    },
                    "slabs": {
                        "description": "Lista de losas (placas) asociadas a esta abscisa. Esta propiedad solo aparecerá si se carga la relación.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Slabs"
                        }
                    },
                    "slabs_with_pathologies": {
                        "description": "Lista de losas (placas) con sus patologías precargadas.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Slabs"
                        }
                    }
                },
                "type": "object"
            },
...
            "Slabs": {
                "title": "Slabs",
                "description": "Modelo de losas de abscisa",
                "required": [
                    "abscisa_id",
                    "long",
                    "width",
                    "type_long",
                    "type_width",
                    "latitude",
                    "longitude"
                ],
                "properties": {
                    "abscisa_id": {
                        "description": "ID de la abscisa",
                        "type": "integer",
                        "example": 1
                    },
                    "long": {
                        "description": "Longitud de la losa",
                        "type": "number",
                        "format": "float",
                        "example": 5
                    },
                    "width": {
                        "description": "Ancho de la losa",
                        "type": "number",
                        "format": "float",
                        "example": 2.5
                    },
                    "type_long": {
                        "description": "Unidad de longitud",
                        "type": "string",
                        "example": "m"
                    },
                    "type_width": {
                        "description": "Unidad de ancho",
                        "type": "string",
                        "example": "m"
                    },
                    "latitude": {
                        "description": "Latitud de la losa",
                        "type": "string",
                        "example": "4.60971"
                    },
                    "longitude": {
                        "description": "Longitud de la losa",
                        "type": "string",
                        "example": "-74.08175"
                    },
                    "observation": {
                        "description": "Observaciones de la losa",
                        "type": "string",
                        "example": "Sin observaciones"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación por Bearer Token (Laravel Sanctum). Incluir el token en el header 'Authorization'.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}